+++ CHANGELOG.md
@@ -9,8 +9,8 @@
 
 ### Fixed issues
 
+- [768](https://github.com/MiKTeX/miktex/issues/768): kpsewhich with option --all
 - [780](https://github.com/MiKTeX/miktex/issues/780): [MikTeX Portable] Can't remove local texmf directory
-
 ## 21.3 - 2021-03-16
 
 ### Upgraded programs